services/client-api/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["simon.islam <simon.islam@brilliant.com.bd>"]
description = ""
license = "GPL-3.0-or-later"
name = "blog-api"
version = "0.1.0"

[tool.poetry.dependencies]
PyJWT = "^2.5.0"
fastapi = "^0.85.0"
gunicorn = "^20.1.0"
motor = "^3.0.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.18.3"}

[tool.poetry.dev-dependencies]
PyYAML = "^6.0"
bandit = "^1.7.4"
black = "^22.8.0"
dlint = "^0.13.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.9.23"
flake8-builtins = "^1.5.3"
isort = "^5.10.1"
openapi-spec-validator = "^0.5.1"
pep8-naming = "^0.13.2"
pre-commit = "^2.20.0"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
safety = "^2.3.1"
schemathesis = "^3.17.4"
toml-sort = "^0.20.1"
tox = "^3.26.0"
